+++ sources/dclfindfile.h
@@ -16,9 +16,9 @@
 extern "C"
 {
 #endif
-int     Dcl_FindFirstFile(char *lpFileName, LPDCL_FIND_DATA lpFindFileData);
-BOOL    Dcl_FindNextFile(int hFindFile, LPDCL_FIND_DATA lpFindFileData);
-BOOL    Dcl_FindClose(int hFindFile);
+intptr_t     Dcl_FindFirstFile(char *lpFileName, LPDCL_FIND_DATA lpFindFileData);
+BOOL    Dcl_FindNextFile(intptr_t hFindFile, LPDCL_FIND_DATA lpFindFileData);
+BOOL    Dcl_FindClose(intptr_t hFindFile);
 #ifdef __cplusplus
 }
 #endif
